audit_tools/defi_attack_vectors.py
"""
DeFi Attack Vector Analysis module for identifying potential attack vectors in DeFi protocols.
"""

from typing import Dict, List, Any
from web3 import Web3
import logging

class DeFiAttackVectorAnalyzer:
    """
    Analyzes potential attack vectors in DeFi protocols.
    """

    def __init__(self, web3_provider: str = None):
        self.logger = logging.getLogger(__name__)
        self.w3 = Web3(Web3.HTTPProvider(web3_provider)) if web3_provider else None

        # Common DeFi attack vectors
        self.attack_vectors = {
            "CRITICAL": {
                "flash_loan_attack": {
                    "pattern": "flash_loan",
                    "description": "Flash loan attack vulnerability",
                    "impact": "Price manipulation and arbitrage exploitation"
                },
                "reentrancy_attack": {
                    "pattern": "reentrancy",
                    "description": "Reentrancy attack vulnerability",
                    "impact": "Unauthorized fund withdrawal and state manipulation"
                },
                "oracle_manipulation": {
                    "pattern": "price_oracle",
                    "description": "Price oracle manipulation vulnerability",
                    "impact": "Price feed manipulation leading to unfair trades"
                }
            },
            "HIGH": {
                "front_running": {
                    "pattern": "front_run",
                    "description": "Front-running vulnerability",
                    "impact": "Transaction ordering exploitation"
                },
                "sandwich_attack": {
                    "pattern": "sandwich",
                    "description": "Sandwich attack vulnerability",
                    "impact": "Price manipulation through surrounding transactions"
                },
                "governance_attack": {
                    "pattern": "governance",
                    "description": "Governance attack vulnerability",
                    "impact": "Malicious proposal execution and token holder manipulation"
                }
            }
        }

    def analyze_attack_vectors(self, contract_address: str) -> Dict[str, Any]:
        """
        Analyze potential attack vectors in a DeFi protocol
        """
        if not self.w3 or not Web3.is_address(contract_address):
            return {"error": "Invalid contract address or web3 provider not configured"}

        findings = {
            "address": contract_address,
            "risk_level": "LOW",
            "attack_vectors": [],
            "mitigations": []
        }

        try:
            # Get contract code
            code = self.w3.eth.get_code(contract_address)
            if code == b'':
                return {"error": "Contract not found or not deployed"}

            # Analyze contract code for attack vectors
            findings.update(self._analyze_vectors(code))

        except Exception as e:
            self.logger.error(f"Error analyzing contract {contract_address}: {str(e)}")
            return {"error": f"Error analyzing contract: {str(e)}"}

        return findings

    def _analyze_vectors(self, contract_code: bytes) -> Dict[str, Any]:
        """
        Analyze contract code for potential attack vectors
        """
        findings = {
            "attack_vectors": [],
            "risk_level": "LOW",
            "mitigations": []
        }

        code_str = contract_code.hex().lower()

        # Check for critical attack vectors
        for vector_name, vector_info in self.attack_vectors["CRITICAL"].items():
            if self._check_vector_pattern(code_str, vector_info["pattern"]):
                findings["attack_vectors"].append({
                    "name": vector_name,
                    "description": vector_info["description"],
                    "impact": vector_info["impact"],
                    "severity": "CRITICAL"
                })
                findings["risk_level"] = "CRITICAL"

        # Check for high-risk attack vectors
        if findings["risk_level"] != "CRITICAL":
            for vector_name, vector_info in self.attack_vectors["HIGH"].items():
                if self._check_vector_pattern(code_str, vector_info["pattern"]):
                    findings["attack_vectors"].append({
                        "name": vector_name,
                        "description": vector_info["description"],
                        "impact": vector_info["impact"],
                        "severity": "HIGH"
                    })
                    findings["risk_level"] = "HIGH"

        # Generate mitigation strategies
        findings["mitigations"] = self._generate_mitigations(findings["attack_vectors"])
        return findings

    def _check_vector_pattern(self, code: str, pattern: str) -> bool:
        """
        Check if an attack vector pattern exists in the contract code
        """
        return pattern.lower().replace("_", "") in code

    def _generate_mitigations(self, vectors: List[Dict[str, str]]) -> List[str]:
        """
        Generate mitigation strategies for identified attack vectors
        """
        mitigations = []
        for vector in vectors:
            if vector["name"] == "flash_loan_attack":
                mitigations.extend([
                    "Implement proper validation for flash loan functions",
                    "Use decentralized price oracles",
                    "Add minimum holding periods for governance tokens"
                ])
            elif vector["name"] == "reentrancy_attack":
                mitigations.extend([
                    "Implement reentrancy guards using OpenZeppelin's ReentrancyGuard",
                    "Follow checks-effects-interactions pattern",
                    "Use pull over push payment patterns"
                ])
            elif vector["name"] == "oracle_manipulation":
                mitigations.extend([
                    "Use multiple price oracles",
                    "Implement proper validation mechanisms",
                    "Add time-weighted average price (TWAP) mechanisms"
                ])
            elif vector["name"] == "front_running":
                mitigations.extend([
                    "Implement commit-reveal schemes",
                    "Use flash-bots to prevent front-running",
                    "Add minimum confirmation blocks"
                ])
            elif vector["name"] == "sandwich_attack":
                mitigations.extend([
                    "Implement slippage protection",
                    "Add minimum output amounts",
                    "Use private transaction pools"
                ])
            elif vector["name"] == "governance_attack":
                mitigations.extend([
                    "Implement timelock mechanisms",
                    "Add proper voting power calculations",
                    "Use quadratic voting mechanisms"
                ])

        return list(set(mitigations))  # Remove duplicates
